feat: migrate from vite SPA to react router v7 framework mode
