+++ .gitignore
@@ -12,6 +12,9 @@ dist
 dist-ssr
 *.local
 
+# React Router
+.react-router/
+
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json
