+++ app/app.css
@@ -111,7 +111,6 @@
     animation: sat-drop 1.2s cubic-bezier(0.16, 1, 0.3, 1) forwards,
                sat-float 6s ease-in-out 1.2s infinite;
   }
-
 }
 
 @keyframes sat-drop {
@@ -133,5 +132,3 @@
   from { opacity: 0; transform: translateY(20px); }
   to { opacity: 1; transform: translateY(0); }
 }
-
-
