+++ app/components/Navbar.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { Link, useLocation } from "react-router-dom";
+import { Link, useLocation } from "react-router";
 import { FaBars } from "react-icons/fa";
 import { IoCloseCircleSharp } from "react-icons/io5";
