+++ app/routes/aboutus-redirect.tsx
@@ -1,0 +1,9 @@
+import { redirect } from "react-router";
+
+export function clientLoader() {
+  return redirect("/aboutus/what-we-do");
+}
+
+export default function AboutUsRedirect() {
+  return null;
+}
