+++ app/routes/index-redirect.tsx
@@ -1,0 +1,9 @@
+import { redirect } from "react-router";
+
+export function clientLoader() {
+  return redirect("/home");
+}
+
+export default function IndexRedirect() {
+  return null;
+}
