--- index.html
@@ -1,42 +1,0 @@
-
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/png" href="/images/CubeSat_logo.png" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta name="theme-color" content="#070A0F" />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap" />
-    <meta name="description" content="UCI CubeSat is a student-led organization at UC Irvine dedicated to designing, building, and launching a 2U nanosatellite into Low Earth Orbit. We pioneer next-generation thermal control technology for space exploration." />
-    <meta name="keywords" content="UCI CubeSat, UC Irvine, nanosatellite, space, aerospace engineering, student organization, thermal control, satellite" />
-    <meta name="author" content="UCI CubeSat Team" />
-    <meta property="og:title" content="UCI CubeSat - Pioneering Next-Gen Thermal Control Technology" />
-    <meta property="og:description" content="A student-led organization at UC Irvine designing and launching a 2U nanosatellite into Low Earth Orbit." />
-    <meta property="og:type" content="website" />
-    <meta property="og:image" content="https://ucicubesat.com/images/og-image.png" />
-    <meta property="og:url" content="https://ucicubesat.com" />
-    <meta name="twitter:card" content="summary_large_image" />
-    <meta name="twitter:title" content="UCI CubeSat - Pioneering Next-Gen Thermal Control Technology" />
-    <meta name="twitter:description" content="A student-led organization at UC Irvine designing and launching a 2U nanosatellite into Low Earth Orbit." />
-    <meta name="twitter:image" content="https://ucicubesat.com/images/og-image.png" />
-    <link rel="canonical" href="https://ucicubesat.com" />
-    <script type="application/ld+json">
-    {
-      "@context": "https://schema.org",
-      "@type": "Organization",
-      "name": "UCI CubeSat",
-      "url": "https://ucicubesat.com",
-      "logo": "https://ucicubesat.com/images/CubeSat_logo.png",
-      "description": "A student-led organization at UC Irvine dedicated to designing, building, and launching a 2U nanosatellite into Low Earth Orbit.",
-      "address": { "@type": "PostalAddress", "addressLocality": "Irvine", "addressRegion": "CA" }
-    }
-    </script>
-    <title>UCI CubeSat</title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
-  </body>
-</html>
