+++ package.json
@@ -4,25 +4,26 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc -b && vite build",
-    "lint": "eslint .",
-    "preview": "vite preview",
-    "deploy": "gh-pages -d dist"
+    "dev": "react-router dev",
+    "build": "react-router build",
+    "preview": "vite preview --outDir dist/client",
+    "typecheck": "react-router typegen && tsc",
+    "lint": "eslint ."
   },
   "dependencies": {
     "@tailwindcss/typography": "^0.5.19",
     "react": "^18.3.1",
     "react-countup": "^6.5.3",
     "react-dom": "^18.3.1",
-    "react-helmet-async": "^2.0.5",
     "react-icons": "^5.5.0",
-    "react-router-dom": "^7.11.0"
+    "@react-router/node": "^7.13.0",
+    "react-router": "^7.13.0",
+    "isbot": "^5"
   },
   "devDependencies": {
+    "@react-router/dev": "^7.13.0",
     "@types/react": "^18.3.27",
     "@types/react-dom": "^18.3.7",
-    "@vitejs/plugin-react": "^4.7.0",
     "autoprefixer": "^10.4.23",
     "eslint": "^9.39.2",
     "eslint-plugin-react-hooks": "^5.2.0",
@@ -31,7 +32,8 @@
     "postcss-nesting": "^13.0.2",
     "tailwindcss": "^3.4.19",
     "typescript": "~5.6.3",
-    "vite": "^6.4.1"
+    "vite": "^6.4.1",
+    "vite-tsconfig-paths": "^5.1.4"
   },
   "optionalDependencies": {
     "@rollup/rollup-linux-x64-gnu": "4.6.1",
