--- src/components/Layout.tsx
@@ -1,21 +1,0 @@
-import { ReactNode } from 'react';
-import Navbar from './Navbar';
-import Footer from './Footer';
-
-interface LayoutProps {
-    children: ReactNode;
-}
-
-const Layout = ({ children }: LayoutProps) => {
-    return (
-        <div className="flex flex-col min-h-screen">
-            <Navbar />
-            <main id="main-content" className="flex-grow">
-                {children}
-            </main>
-            <Footer />
-        </div>
-    );
-};
-
-export default Layout;
