--- src/hooks/useCountUp.ts
@@ -1,67 +1,0 @@
-import { useEffect, useState, useRef } from 'react';
-
-interface UseCountUpOptions {
-  end: number;
-  duration?: number;
-  prefix?: string;
-  suffix?: string;
-}
-
-export function useCountUp({ end, duration = 2000, prefix = '', suffix = '' }: UseCountUpOptions) {
-  const [count, setCount] = useState(0);
-  const [hasAnimated, setHasAnimated] = useState(false);
-  const ref = useRef<HTMLElement>(null);
-
-  useEffect(() => {
-    const observer = new IntersectionObserver(
-      ([entry]) => {
-        if (entry.isIntersecting && !hasAnimated) {
-          setHasAnimated(true);
-          
-          const startTime = performance.now();
-          const startValue = 0;
-          
-          const animate = (currentTime: number) => {
-            const elapsed = currentTime - startTime;
-            const progress = Math.min(elapsed / duration, 1);
-            
-            // Easing function (ease-out)
-            const easeOut = 1 - Math.pow(1 - progress, 3);
-            const currentValue = Math.floor(startValue + (end - startValue) * easeOut);
-            
-            setCount(currentValue);
-            
-            if (progress < 1) {
-              requestAnimationFrame(animate);
-            } else {
-              setCount(end);
-            }
-          };
-          
-          requestAnimationFrame(animate);
-        }
-      },
-      {
-        threshold: 0.3,
-        rootMargin: '0px',
-      }
-    );
-
-    if (ref.current) {
-      observer.observe(ref.current);
-    }
-
-    return () => {
-      if (ref.current) {
-        observer.unobserve(ref.current);
-      }
-    };
-  }, [end, duration, hasAnimated]);
-
-  const displayValue = `${prefix}${count}${suffix}`;
-  
-  return { ref, displayValue };
-}
-
-
-
