--- src/pages/aboutus/AboutUs.tsx
@@ -1,8 +1,0 @@
-export default function AboutUs() {
-    return (
-        <div className="p-8 max-w-[1200px] mx-auto">
-            <h1 className="text-5xl font-bold mb-4 text-center text-primary">About Us</h1>
-            <p className="text-lg leading-relaxed text-center">Coming soon...</p>
-        </div>
-    );
-}
