--- src/pages/aboutus/meet-the-team/MeetTheTeam.tsx
@@ -1,310 +1,0 @@
-import { useState } from 'react';
-import { Helmet } from 'react-helmet-async';
-import { TeamMakeup } from '../../../components/TeamMakeup/TeamMakeup';
-import { useScrollAnimation } from '../../../hooks/useScrollAnimation';
-
-const MeetTheTeam = () => {
-  const [activeRole, setActiveRole] = useState("All");
-  const headerAnimation = useScrollAnimation<HTMLDivElement>();
-  const gridAnimation = useScrollAnimation<HTMLDivElement>();
-
-  const teamMembers = [
-    {
-      name: "Sonia Cruze",
-      role: "Executive",
-      position: "Project Manager",
-      image: "/Headshots/SoniaCruze.jpg",
-    },
-    {
-      name: "Byren Cheema",
-      role: "Advisory",
-      position: "Webmaster",
-      image: "",
-    },
-    {
-      name: "Cosmo Daniels",
-      role: "Avionics",
-      position: "Hardware Lead, Software Lead",
-      image: "",
-    },
-    {
-      name: "Enzo Li",
-      role: "Avionics",
-      position: "Hardware",
-      image: "",
-    },
-    {
-      name: "Justin Cabrera",
-      role: "Avionics",
-      position: "Hardware",
-      image: "/Headshots/JustinCabrera.jpg",
-    },
-    {
-      name: "Rohit De",
-      role: "Avionics",
-      position: "Hardware",
-      image: "/Headshots/RohitDe.jpg",
-    },
-    {
-      name: "Mahek Ali",
-      role: "Avionics",
-      position: "Hardware",
-      image: "/Headshots/MahekAli.jpg",
-    },
-    {
-      name: "Phyo Pyae Moe Aung",
-      role: "Avionics",
-      position: "Software",
-      image: "",
-    },
-    {
-      name: "Tanvi Akula",
-      role: "Avionics",
-      position: "Software",
-      image: "/Headshots/TanviAkula.jpg",
-    },
-    {
-      name: "Urja Vaidya",
-      role: "Avionics",
-      position: "Software",
-      image: "/Headshots/UrjaVaidya.jpg",
-    },
-    {
-      name: "Emmanuel Bivian",
-      role: "Avionics",
-      position: "Software",
-      image: "",
-    },
-    {
-      name: "Hailey Williams",
-      role: "Avionics",
-      position: "Software",
-      image: "/Headshots/HaileyWilliams.jpg",
-    },
-    {
-      name: "Jonathan Lin",
-      role: "Communications",
-      position: "Subteam Lead",
-      image: "/Headshots/JonathanLin.jpg",
-    },
-    {
-      name: "Raiche Gutierrez",
-      role: "Communications",
-      position: "Hardware",
-      image: "/Headshots/RaicheGutierrez.jpg",
-    },
-    {
-      name: "An Bui",
-      role: "Communications",
-      position: "Hardware",
-      image: "/Headshots/AnBui.jpg",
-    },
-    {
-      name: "Ibrahim Hussein",
-      role: "Communications",
-      position: "Hardware",
-      image: "/Headshots/IbrahimHussein.jpg",
-    },
-    {
-      name: "Steven Yu",
-      role: "Communications",
-      position: "Hardware",
-      image: "/Headshots/StevenYu.jpg",
-    },
-    {
-      name: "Bart Backer",
-      role: "Communications",
-      position: "Hardware",
-      image: "/Headshots/BartBacker.jpg",
-    },
-    {
-      name: "Derick Barrientos",
-      role: "Communications",
-      position: "Software",
-      image: "/Headshots/DerickBarrientos.jpg",
-    },
-    {
-      name: "Alejandro Gomez",
-      role: "Communications",
-      position: "Software",
-      image: "",
-    },
-    {
-      name: "Jenavieve Steward",
-      role: "Power",
-      position: "Subteam Lead",
-      image: "/Headshots/JenavieveSteward.jpg",
-    },
-    {
-      name: "Akam Khinda",
-      role: "Power",
-      position: "Hardware",
-      image: "",
-    },
-    {
-      name: "Eric Yan",
-      role: "Power",
-      position: "Hardware",
-      image: "/Headshots/EricYan.jpg",
-    },
-    {
-      name: "Andy Li",
-      role: "Power",
-      position: "Hardware",
-      image: "",
-    },
-    {
-      name: "Alex Ikeda",
-      role: "Power",
-      position: "Engineer, Hardware",
-      image: "/Headshots/AlexIkeda.jpg",
-    },
-    {
-      name: "Sijia Zhang",
-      role: "Power",
-      position: "Engineer",
-      image: "",
-    },
-    {
-      name: "Nathan Kwock",
-      role: "Power",
-      position: "Software",
-      image: "",
-    },
-    {
-      name: "Sofia Fowler",
-      role: "Power",
-      position: "Engineer, Software",
-      image: "",
-    },
-    {
-      name: "Gloria McMaster-Sanchez",
-      role: "Structures",
-      position: "Subteam Lead",
-      image: "/Headshots/GloriaMcmasterSanchez.jpg",
-    },
-    {
-      name: "Supreya Saxena",
-      role: "Structures",
-      position: "Engineer",
-      image: "/Headshots/SupreyaSaxena.jpg",
-    },
-    {
-      name: "Yuvraj Chera",
-      role: "Structures",
-      position: "Engineer",
-      image: "/Headshots/YuvrajChera.jpg",
-    },
-    {
-      name: "Nora Shah",
-      role: "Structures",
-      position: "Engineer",
-      image: "/Headshots/NoraShah.jpg",
-    },
-    {
-      name: "Joshua Luna",
-      role: "Structures",
-      position: "Engineer",
-      image: "/Headshots/JoshuaLuna.jpg",
-    },
-    {
-      name: "Sarghi Kamboj",
-      role: "Structures",
-      position: "Engineer",
-      image: "/Headshots/SarghiKamboj.jpg",
-    },
-    {
-      name: "Xavier Marciano",
-      role: "Systems",
-      position: "Subteam Lead, Integration and Design",
-      image: "",
-    },
-    {
-      name: "Briant Fernandez",
-      role: "Systems",
-      position: "Engineer",
-      image: "",
-    },
-    {
-      name: "Katie Lien",
-      role: "Systems",
-      position: "Integration and Design",
-      image: "",
-    },
-    {
-      name: "Brayden Weimholt",
-      role: "Systems",
-      position: "Software",
-      image: "/Headshots/BraydenWeimholt.jpg",
-    },
-  ];
-
-  const filteredMembers = activeRole === "All"
-    ? teamMembers
-    : teamMembers.filter(member => member.role === activeRole);
-
-  return (
-    <>
-      <Helmet>
-        <title>Meet the Team | UCI CubeSat</title>
-        <meta name="description" content="Meet the interdisciplinary team of undergraduate engineers at UCI CubeSat building the future of space technology." />
-        <link rel="canonical" href="https://ucicubesat.com/aboutus/meet-the-team" />
-        <meta property="og:title" content="Meet the Team | UCI CubeSat" />
-        <meta property="og:description" content="Meet the interdisciplinary team of undergraduate engineers at UCI CubeSat." />
-        <meta property="og:url" content="https://ucicubesat.com/aboutus/meet-the-team" />
-        <meta name="twitter:card" content="summary_large_image" />
-        <meta name="twitter:title" content="Meet the Team | UCI CubeSat" />
-        <meta name="twitter:description" content="Meet the interdisciplinary team of undergraduate engineers at UCI CubeSat." />
-        <meta name="twitter:image" content="https://ucicubesat.com/images/og-image.png" />
-      </Helmet>
-      <div className="flex flex-col items-center w-full px-6">
-        <div
-          ref={headerAnimation.ref}
-          className={`pt-32 pb-10 text-center animate-on-scroll ${headerAnimation.isVisible ? 'visible' : ''}`}
-        >
-          <p className="text-xs font-medium tracking-[0.15em] uppercase text-earth mb-4 max-sm:text-[11px]">
-            Our People
-          </p>
-          <h1 className="font-bold text-[48px] text-primary m-0 mb-3 max-sm:text-[32px]">
-            Meet Our Team.
-          </h1>
-          <p className="text-[15px] text-muted max-w-[480px] mx-auto">
-            An interdisciplinary group of undergraduate engineers building the future of space technology.
-          </p>
-        </div>
-        <TeamMakeup
-          teamMembers={teamMembers}
-          activeRole={activeRole}
-          onRoleChange={setActiveRole}
-        />
-        <div
-          ref={gridAnimation.ref}
-          className={`relative section-glow-earth pb-16 animate-on-scroll ${gridAnimation.isVisible ? 'visible' : ''}`}
-        >
-          <div className="grid grid-cols-[repeat(auto-fit,minmax(200px,1fr))] gap-8 w-full max-w-[1200px] max-sm:grid-cols-[repeat(auto-fit,minmax(150px,1fr))] max-sm:gap-6">
-            {filteredMembers.map((member, index) => (
-              <div className="flex flex-col items-center text-center" key={index}>
-                <div className="w-[120px] h-[120px] bg-orbital rounded-full flex justify-center items-center mb-4 overflow-hidden border-2 border-starlight">
-                  {member.image ? (
-                    <img src={member.image} alt={member.name} width={120} height={120} loading="lazy" className="w-full h-full object-cover" />
-                  ) : (
-                    <div className="text-[28px] font-medium text-earth tracking-[2px]">
-                      {member.name.split(' ').map(n => n[0]).join('')}
-                    </div>
-                  )}
-                </div>
-                <div className="text-sm">
-                  <h3 className="my-1 text-sm font-medium text-muted">{member.role}</h3>
-                  <p className="text-earth my-1 text-[13px] font-medium">{member.position}</p>
-                  <span className="text-primary font-medium block my-1">{member.name}</span>
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      </div>
-    </>
-  );
-};
-
-export default MeetTheTeam;
